backend/requirements.txt
asgiref==3.11.0
Django==5.2.10
djangorestframework==3.16.1
psycopg2-binary
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.15.0
gunicorn==20.1.0